Agencia OK uso de BigNumeral em Conta e começo de Transação

+++ src/main/java/acc/br/accenturebank/dto/ContaDetalhesDTO.java
@@ -4,11 +4,13 @@ package acc.br.accenturebank.dto;
 import acc.br.accenturebank.model.enums.TipoConta;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ContaDetalhesDTO {
     private int idConta;
     private String numero;
-    private float saldo;
+    private BigDecimal saldo;
     private boolean ativa;
     private boolean pixAtivo;
     private String chavePix;
