+++ src/main/java/acc/br/accenturebank/dto/ContaResponseDTO.java
@@ -4,11 +4,13 @@ import acc.br.accenturebank.model.Agencia;
 import acc.br.accenturebank.model.enums.TipoConta;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ContaResponseDTO {
     private int idConta;
     private String numero;
-    private float saldo;
+    private BigDecimal saldo;
     private boolean ativa;
     private boolean pixAtivo;
     private String chavePix;
