+++ src/main/java/acc/br/accenturebank/dto/RecargaCelularRequest.java
@@ -2,8 +2,10 @@ package acc.br.accenturebank.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class RecargaCelularRequest {
     private String numeroCelular;
-    private float valor;
+    private BigDecimal valor;
 }
